creating remove from favorites funcionality

+++ client/src/contexts/CartContext.js
@@ -1,5 +1,5 @@
 import axios from "axios";
-import { useContext, createContext ,useState, useEffect } from "react";
+import { useContext, createContext, useState, useEffect } from "react";
 
 const CartContext = createContext()
 
@@ -7,45 +7,68 @@ export const useCart = () => {
     return useContext(CartContext)
 }
 
-export const CartProvider = ({children}) => {
+export const CartProvider = ({ children }) => {
 
-    const [favorites,setFavorites] = useState(JSON.parse(localStorage.getItem('favorites')) || [])
+    const [favorites, setFavorites] = useState(JSON.parse(localStorage.getItem('favorites')) || [])
 
 
 
     useEffect(() => {
-        console.log(favorites) 
-    },[])
+        console.log(favorites)
+    }, [])
+
+
     const addToFavorites = (product) => {
 
-        const {productName, photoOfSeller,price,url,image} = product
+        const { productName, photoOfSeller, price, url, image } = product
+        const id = localStorage.getItem('id')
+        let exist = favorites.find(x => x.url === product.url);
 
-        axios.post('https://shopnet.up.railway.app/api/users/addToFavorites', {
-            id: localStorage.getItem('id'),
-            productName,
-            photoOfSeller,
-            price,
-            url,
-            image
-        })
-        .then( res => {
-            localStorage.setItem('favorites', JSON.stringify(res.data.updated.favorites))
-            setFavorites([...favorites, {photoOfSeller,productName,price,url,image}])
+
+        if (!exist) {
+            setFavorites([...favorites, { photoOfSeller, productName, price, url, image }])
+            axios.post('https://shopnet.up.railway.app/api/users/addToFavorites', {
+                id,
+                productName,
+                photoOfSeller,
+                price,
+                url,
+                image
+            })
+                .then(res => {
+                    localStorage.setItem('favorites', JSON.stringify(res.data.updated.favorites))
+                }
+                )
+                .catch(e => console.log(e))
         }
-        )
-        .catch(e => console.log(e))
-}
 
 
-const value = {
-    addToFavorites,
-    favorites
-}
+        if (exist) {
+            const newFavorites = favorites.filter(x => x.url !== url)
+            console.log(newFavorites)
+            setFavorites(newFavorites)
+            axios.post('https://shopnet.up.railway.app/api/users/removeFromFavorites', {
+                id,
+                url
+            })
+                .then((res) => {
+                    localStorage.setItem('favorites', JSON.stringify(res.data.updated.favorites))
+                })
+                .catch(e => console.log(e))
+        }
+    }
 
-return (
-    <CartContext.Provider value={value}>
-        {children}
-    </CartContext.Provider>
-)
+
+
+    const value = {
+        addToFavorites,
+        favorites
+    }
+
+    return (
+        <CartContext.Provider value={value}>
+            {children}
+        </CartContext.Provider>
+    )
 }
